llm_backend/requirements.txt
fastapi
uvicorn
jinja2
requests
python-dotenv
openai
fastmcp